Refactor combat tab into subcomponents
